Fixed Line Spacing and Font Size Selection
Fixed the issue of line spacing in the rte, as well as an issue I found that occurs when you select text with more than 1 font size. It would pop up a dependency unset value text in the font size combobox. It should no longer show that and instead just be blank.